Split Processor, now labeler & extractor, updated README and Docstrings
