+++ aEye/__init__.py
@@ -1,2 +1,4 @@
 from aEye.video import Video
-from aEye.processor import Processor
+from aEye.labeler import Labeler
+from aEye.extractor import Extractor
+from aEye.auxiliary import Aux
