+++ requirements.txt
@@ -2,4 +2,5 @@ boto3>=1.26.29
 opencv-python>=4.7.0.00
 numpy>=1.21.6
 static-ffmpeg>=2.5
+opencv-python-headless>=4.8.0.00
